- introduce `InlineHyperZone`, which takes care of inlining prefixed props to symbolic props, before passing them onto the renderers. - move `/src/hyperzone.ts` to `/src/hzone/renderers.ts`, as `HyperZoneRender` can be thought of as a framework itself, no?
