+++ src/mod.ts
@@ -1,6 +1,6 @@
-export * from "./vanilla/mod.ts"
 export { normalizeAttrProps, stringify, stringifyAttrValue } from "./funcdefs.ts"
-export { HyperZone } from "./hyperzone.ts"
+export { HyperZone, InlineHyperZone } from "./hzone/mod.ts"
 export type * from "./typedefs.ts"
 export { ADVANCED_EVENTS, ATTRS, EVENTS, EXECUTE, Fragment, HyperRender, MEMBERS } from "./typedefs.ts"
+export * from "./vanilla/mod.ts"
