+++ src/tsignal_inline/renderers.ts
